Feat: add main route screen
