+++ build.gradle.kts
@@ -8,4 +8,3 @@ plugins {
     alias(libs.plugins.hilt.android) apply false
     alias(libs.plugins.ksp) apply false
 }
-
