+++ feature/camera/.gitignore
