+++ feature/camera/proguard-rules.pro
