+++ settings.gradle.kts
@@ -25,5 +25,9 @@ enableFeaturePreview("TYPESAFE_PROJECT_ACCESSORS")
 include(":app")
 include(":core:model")
 include(":core:data")
-include(":feature:main")
 include(":core:domain")
+include(":feature:home")
+include(":feature:follower")
+include(":feature:my")
+include(":feature:search")
+include(":feature:camera")
